:construction: build dashboard page [over 15 hours of work in 1 commit]
I kinda sat for about 15 hours and wrote the code. The problem it took that long was because I had to make sure everything was responsive It was specially difficult because I used my own framework, which is still a little buggy. Learnt a lot about the functionalities of Grid so it was a good thing in the end. Hope to complete it in about 2-3 hours of work, then move on to the other pages, which shouldn't take much longer as the form is similar on the other pages.